Rework database code into smaller classes.
Not entirely sure I want to keep all of this code - I feel like Rails should own some of the smarts itself - but there's a good chance not every version of Rails behaves quite the same way, and even less of a chance that Rails' internals for managing databases can be called reliably across many versions. So, sticking with this for the moment.