Have a reasonable fallback for unknown modules.
Mostly, this code won't be hit - but it allows for explicit support of new Rails components if/when others are added in later. This used to be the logic in play (when it was presumed every component had a railtie - not the case for ActiveStorage, mind you), and the given keys were just appended with /railtie. Now, that's just the default behaviour, but known libraries are added at the specific Rails versions they were introduced in.